"Textual 7" -> "Textual"
Not that I plan on a version bump. Just that having the 7 in the name is a little too verbose at this point. I do not like it.

This commit includes misc. other changes that interface builder applies to xib files when editing that I couldn't care to exclude. They are harmless.