Fix issue that may cause outgoing commands to be sent twice
The function for sending commands has an option to bypass flood control queue and go directly to socket. This is only used for PONG as of this commit which is why this issue has gone unnoticed for so long. When the flag bypassQueue was true, the PONG was being sent, but the function did not return afterwards, so the PONG was queued for another send afterwards.