Modify TPCPreferencesUserDefaults
All preferences related property lists have been moved to a subdirectory in the Resources folder of the app.
Keys not allowed to be exported has been removed from StaticStore.plist and place in its own file.
Keys not allowed to be migrated has been added.
Keys that are obsolete have been removed from StaticStore.plist. No replacement.
A master list of all known preference keys, including those that are hidden or used internally, has been added.
When asking if a key is safe to export, it will first check the list of forbidden keys. It will then check if the key appears in the master list.
Same logic applies when asking if a key can be migrated.
An integer based comparator is now defined for each key instead of a string.