Modify scrollback manager
When initializing scrollback manager, pass directory in which to save database. No longer pass complete file path. The service will manage the filename so that it can changed it if need be.

When there is a need for the service to change the name of the database, such as being unable to open it, then it will no longer attempt to delete the old copy as well.