Modify Server Properties
Notify user when there are no certificates that can be used for identifying.
Disabled selection of certificate fingerprints. Selecting these text fields make the contents grow beyond the column. Instead of debugging that behavior, it is easier to just disable selection. There is already a copy button associated with each.